- Add fixes for iOS simulator, it now runs in the simulator  - Add lots of XCode attributes  - Fix CMake generator used for iOS builds  - Add new storage type, CachedFile, not deleted 100% of the time like TemporaryFile  - Redirect config files, temporary files and cached files to iOS directories  - Fix iOS requirement of armv7 and armv8, it broke simulator  - Fix plist for iOS, wrong executable name  - Temporarily fix mega-warnings for types in simulator

+++ include/coffee/core/plat/timing/unix/timers.h
@@ -97,7 +97,7 @@ struct UnixTime : PosixIshTimeDef
         CString out;
         out.resize(100);
 
-        int len = std::strftime(&out[0],out.size(),fmt,&t);
+        size_t len = std::strftime(&out[0],out.size(),fmt,&t);
         out.resize(len);
 
         return out;
